wip: window compaction task
